+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Added String#dissect(open, close)
 * Added RegExp.interpret(pattern)
 * Added RegExp#getPattern()
+* Added RegExp#getFlags()
 
 ## 0.1.1 (2014-06-23)